Closes #19
Add ShowAndHidePasswordFilledField
