+++ frontend/package.json
@@ -15,10 +15,10 @@
     "@emotion/styled": "^11.10.4",
     "@mui/icons-material": "^5.10.6",
     "@mui/material": "^5.10.8",
+    "@vitejs/plugin-react-refresh": "^1.3.1",
     "react": "^17.0.0",
     "react-dom": "^17.0.0",
     "react-router-dom": "^6.4.1",
-    "@vitejs/plugin-react-refresh": "^1.3.1",
     "vite": "^2.5.2"
   },
   "devDependencies": {
