+++ frontend/src/components/ShowAndHidePasswordFilledField/ShowAndHidePasswordFilledField.tsx
@@ -1,0 +1,79 @@
+import * as React from "react";
+import FormControl from "@mui/material/FormControl";
+import Visibility from "@mui/icons-material/Visibility";
+import VisibilityOff from "@mui/icons-material/VisibilityOff";
+import InputAdornment from "@mui/material/InputAdornment";
+import InputLabel from "@mui/material/InputLabel";
+import IconButton from "@mui/material/IconButton";
+import FilledInput from "@mui/material/FilledInput";
+import { SxProps } from "@mui/material";
+
+interface State {
+  password: string;
+  showPassword: boolean;
+}
+
+type Props = {
+  label?: string;
+  sx?: SxProps;
+};
+
+const DefaultProps = {
+  label: "Password",
+  sx: {},
+};
+
+export default function ShowAndHidePasswordFilledField({
+  label,
+  sx,
+}: Props): JSX.Element {
+  const [values, setValues] = React.useState<State>({
+    password: "",
+    showPassword: false,
+  });
+
+  const handleChange =
+    (prop: keyof State) => (event: React.ChangeEvent<HTMLInputElement>) => {
+      setValues({ ...values, [prop]: event.target.value });
+    };
+
+  const handleClickShowPassword = () => {
+    setValues({
+      ...values,
+      showPassword: !values.showPassword,
+    });
+  };
+
+  const handleMouseDownPassword = (
+    event: React.MouseEvent<HTMLButtonElement>
+  ) => {
+    event.preventDefault();
+  };
+
+  return (
+    <FormControl required fullWidth variant="filled" sx={sx}>
+      <InputLabel htmlFor="filled-adornment-password">{label}</InputLabel>
+      <FilledInput
+        id="password"
+        name="password"
+        type={values.showPassword ? "text" : "password"}
+        value={values.password}
+        onChange={handleChange("password")}
+        endAdornment={
+          <InputAdornment position="end">
+            <IconButton
+              aria-label="toggle password visibility"
+              onClick={handleClickShowPassword}
+              onMouseDown={handleMouseDownPassword}
+              edge="end"
+            >
+              {values.showPassword ? <VisibilityOff /> : <Visibility />}
+            </IconButton>
+          </InputAdornment>
+        }
+      />
+    </FormControl>
+  );
+}
+
+ShowAndHidePasswordFilledField.defaultProps = DefaultProps;
