+++ frontend/src/pages/Login/LoginForm.tsx
@@ -2,7 +2,6 @@ import * as React from "react";
 import Button from "@mui/material/Button";
 import CssBaseline from "@mui/material/CssBaseline";
 import TextField from "@mui/material/TextField";
-import FormControl from "@mui/material/FormControl";
 import FormControlLabel from "@mui/material/FormControlLabel";
 import Checkbox from "@mui/material/Checkbox";
 import Link from "@mui/material/Link";
@@ -10,12 +9,6 @@ import Grid from "@mui/material/Grid";
 import Box from "@mui/material/Box";
 import Typography from "@mui/material/Typography";
 import Container from "@mui/material/Container";
-import Visibility from "@mui/icons-material/Visibility";
-import VisibilityOff from "@mui/icons-material/VisibilityOff";
-import InputAdornment from "@mui/material/InputAdornment";
-import InputLabel from "@mui/material/InputLabel";
-import IconButton from "@mui/material/IconButton";
-import FilledInput from "@mui/material/FilledInput";
 
 import logo from "../../assets/logo.png";
 import googleLogo from "../../assets/googleLogo.svg";
@@ -23,36 +16,9 @@ import githubLogo from "../../assets/githubLogo.svg";
 import facebookLogo from "../../assets/facebookLogo.svg";
 import twitterLogo from "../../assets/twitterLogo.svg";
 import Colors from "../../styles/Colors";
-
-interface State {
-  password: string;
-  showPassword: boolean;
-}
+import ShowAndHidePasswordFilledField from "../../components/ShowAndHidePasswordFilledField/ShowAndHidePasswordFilledField";
 
 export default function LoginForm(): JSX.Element {
-  const [values, setValues] = React.useState<State>({
-    password: "",
-    showPassword: false,
-  });
-
-  const handleChange =
-    (prop: keyof State) => (event: React.ChangeEvent<HTMLInputElement>) => {
-      setValues({ ...values, [prop]: event.target.value });
-    };
-
-  const handleClickShowPassword = () => {
-    setValues({
-      ...values,
-      showPassword: !values.showPassword,
-    });
-  };
-
-  const handleMouseDownPassword = (
-    event: React.MouseEvent<HTMLButtonElement>
-  ) => {
-    event.preventDefault();
-  };
-
   const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
   };
@@ -98,30 +64,7 @@ export default function LoginForm(): JSX.Element {
             variant="filled"
             sx={{ mb: 3 }}
           />
-          <FormControl required fullWidth variant="filled">
-            <InputLabel htmlFor="filled-adornment-password">
-              Password
-            </InputLabel>
-            <FilledInput
-              id="password"
-              name="password"
-              type={values.showPassword ? "text" : "password"}
-              value={values.password}
-              onChange={handleChange("password")}
-              endAdornment={
-                <InputAdornment position="end">
-                  <IconButton
-                    aria-label="toggle password visibility"
-                    onClick={handleClickShowPassword}
-                    onMouseDown={handleMouseDownPassword}
-                    edge="end"
-                  >
-                    {values.showPassword ? <VisibilityOff /> : <Visibility />}
-                  </IconButton>
-                </InputAdornment>
-              }
-            />
-          </FormControl>
+          <ShowAndHidePasswordFilledField />
           <FormControlLabel
             control={<Checkbox value="remember" color="primary" />}
             label="Remember me"
